# api/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "AlbumDetail": {
        "properties": {
          "artist": {
            "default": "",
            "title": "Artist",
            "type": "string"
          },
          "artist_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "default": "",
            "title": "Artist Id"
          },
          "cover": {
            "default": "",
            "title": "Cover",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "nb_tracks": {
            "default": 0,
            "title": "Nb Tracks",
            "type": "integer"
          },
          "release_date": {
            "default": "",
            "title": "Release Date",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "id"
        ],
        "title": "AlbumDetail",
        "type": "object"
      },
      "AlbumSummary": {
        "properties": {
          "artist": {
            "default": "",
            "title": "Artist",
            "type": "string"
          },
          "cover": {
            "default": "",
            "title": "Cover",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "release_date": {
            "default": "",
            "title": "Release Date",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "AlbumSummary",
        "type": "object"
      },
...
      "ArtistAbout": {
        "properties": {
          "bio": {
            "default": "",
            "title": "Bio",
            "type": "string"
          },
          "listeners": {
            "default": 0,
            "title": "Listeners",
            "type": "integer"
          },
          "playcount": {
            "default": 0,
            "title": "Playcount",
            "type": "integer"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "title": "ArtistAbout",
        "type": "object"
      },
      "ArtistDetail": {
        "properties": {
          "albums": {
            "items": {
              "$ref": "#/components/schemas/AlbumSummary"
            },
            "title": "Albums",
            "type": "array"
          },
          "albums_count": {
            "default": 0,
            "title": "Albums Count",
            "type": "integer"
          },
          "fans": {
            "default": 0,
            "title": "Fans",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "picture": {
            "default": "",
            "title": "Picture",
            "type": "string"
          },
          "related": {
            "items": {
              "$ref": "#/components/schemas/ArtistSummary"
            },
            "title": "Related",
            "type": "array"
          },
          "top": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "title": "Top",
            "type": "array"
          }
        },
        "required": [
          "id"
        ],
        "title": "ArtistDetail",
        "type": "object"
      },
      "ArtistRef": {
        "description": "A single performer on a track (a track can have several).",
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "default": "",
            "title": "Id"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          }
        },
        "title": "ArtistRef",
        "type": "object"
      },
      "ArtistSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "picture": {
            "default": "",
            "title": "Picture",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "ArtistSummary",
        "type": "object"
      },
      "AuthError": {
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "GenreDetail": {
        "properties": {
          "albums": {
            "items": {
              "$ref": "#/components/schemas/AlbumSummary"
            },
            "title": "Albums",
            "type": "array"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistSummary"
            },
            "title": "Artists",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "picture": {
            "default": "",
            "title": "Picture",
            "type": "string"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "id"
        ],
        "title": "GenreDetail",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlayQuery": {
        "properties": {
          "artist": {
            "default": "",
            "title": "Artist",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "PlayQuery",
        "type": "object"
      },
...
      "PlaylistSearchResult": {
        "properties": {
          "cover": {
            "default": "",
            "title": "Cover",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "track_count": {
            "default": 0,
            "title": "Track Count",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "PlaylistSearchResult",
        "type": "object"
      },
...
      "PlaysRequest": {
        "properties": {
          "tracks": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlayQuery"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "title": "PlaysRequest",
        "type": "object"
      },
...
      "Track": {
        "properties": {
          "album": {
            "default": "",
            "title": "Album",
            "type": "string"
          },
          "album_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "default": "",
            "title": "Album Id"
          },
          "artist": {
            "default": "",
            "title": "Artist",
            "type": "string"
          },
          "artist_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "default": "",
            "title": "Artist Id"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistRef"
            },
            "title": "Artists",
            "type": "array"
          },
          "cover": {
            "default": "",
            "title": "Cover",
            "type": "string"
          },
          "duration_sec": {
            "default": 0,
            "title": "Duration Sec",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "rank": {
            "default": 0,
            "title": "Rank",
            "type": "integer"
          },
          "release_date": {
            "default": "",
            "title": "Release Date",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "Track",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sf_session": {
        "description": "HttpOnly JWT session cookie set by the login or registration endpoint.",
        "in": "cookie",
        "name": "sf_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "LoggeRythm API",
    "version": "1.0.0",
    "x-loggerythm-compatible-contract-versions": [
      "v1"
    ],
    "x-loggerythm-contract-version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/albums/{album_id}": {
      "get": {
        "operationId": "album_api_albums__album_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "album_id",
            "required": true,
            "schema": {
              "title": "Album Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Album",
        "tags": [
          "browse"
        ]
      }
    },
    "/api/artist-about": {
      "get": {
        "description": "Last.fm artist biography + stats for the \"About\" section (lazy-loaded).",
        "operationId": "artist_about_api_artist_about_get",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "default": "",
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistAbout"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artist About",
        "tags": [
          "browse"
        ]
      }
    },
    "/api/artists/{artist_id}": {
      "get": {
        "operationId": "artist_api_artists__artist_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artist_id",
            "required": true,
            "schema": {
              "title": "Artist Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artist",
        "tags": [
          "browse"
        ]
      }
    },
...
    "/api/deezer-playlist/{playlist_id}": {
      "get": {
        "operationId": "deezer_playlist_api_deezer_playlist__playlist_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "title": "Playlist Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Deezer Playlist Api Deezer Playlist  Playlist Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deezer Playlist",
        "tags": [
          "browse"
        ]
      }
    },
...
    "/api/genres/{genre_id}": {
      "get": {
        "operationId": "genre_api_genres__genre_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "genre_id",
            "required": true,
            "schema": {
              "title": "Genre Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenreDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Genre",
        "tags": [
          "browse"
        ]
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "track",
              "title": "Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Track"
                  },
                  "title": "Response Search Api Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "browse"
        ]
      }
    },
    "/api/search/artist": {
      "get": {
        "operationId": "search_artist_api_search_artist_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtistSummary"
                  },
                  "title": "Response Search Artist Api Search Artist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Artist",
        "tags": [
          "browse"
        ]
      }
    },
    "/api/search/playlist": {
      "get": {
        "operationId": "search_playlist_api_search_playlist_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaylistSearchResult"
                  },
                  "title": "Response Search Playlist Api Search Playlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Playlist",
        "tags": [
          "browse"
        ]
      }
    },
    "/api/track-plays": {
      "post": {
        "description": "Batched Last.fm play counts for a set of tracks → ``{id: {plays, listeners}}``.",
        "operationId": "track_plays_api_track_plays_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaysRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Track Plays Api Track Plays Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Track Plays",
        "tags": [
          "browse"
        ]
      }
    },
    "/api/tracks/{deezer_id}": {
      "get": {
        "operationId": "track_metadata_api_tracks__deezer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deezer_id",
            "required": true,
            "schema": {
              "title": "Deezer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Track"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track Metadata",
        "tags": [
          "browse"
        ]
      }
    },
...
  }
}